fix warn for clang++13

+++ src/topling/block_based_table_side_plugin.cc
@@ -21,7 +21,7 @@ void CleanSpace(std::string& s) {
 // s contains some keys and values
 json SliceStringIntoJson(std::string s) {
   json rtn;
-  int pre_index = 0;
+  size_t pre_index = 0;
   std::string temp_key;
 
   CleanSpace(s);
